refactor(writer): deduplicate CSS, clean inline styles, fix resizer bug
- Extract 30+ duplicate CSS links into shared _writer_css.html partial
- Replace emoji icons (moon/sun) with Font Awesome icons
- Remove inline styles from templates, move to CSS classes
- Remove dead #compilation-panel code (105 lines)
- Simplify fallback panel switcher (73 → 14 lines)
- Remove unnecessary file-load toast notification
- Fix details panel resizer: add htmx:afterSettle hook for late-init
- Split oversized Python files (demo_plot: 1505→89, arxiv_categories: 593→59)
