+++ apps/vis_app/static/vis_app/img/plot_gallery/__init__.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""Plot gallery package - publication-ready demo plots."""
+
+# EOF
