+++ apps/writer_app/management/commands/arxiv_categories_data.py
@@ -4,590 +4,55 @@
 
 DEPRECATED: arXiv integration moved to separate service.
 This data is kept for reference only.
+
+Category data is stored in arxiv_categories.json alongside this file.
 """
 
 from __future__ import annotations
 
-# Computer Science categories
-CS_CATEGORIES = [
-    {
-        "code": "cs.AI",
-        "name": "Artificial Intelligence",
-        "description": "Covers all areas of AI except Vision, Robotics, Machine Learning, Multiagent Systems, and Computation and Language.",
-    },
-    {
-        "code": "cs.AR",
-        "name": "Hardware Architecture",
-        "description": "Covers systems organization and hardware architecture.",
-    },
-    {
-        "code": "cs.CC",
-        "name": "Computational Complexity",
-        "description": "Covers models of computation, complexity classes, structural complexity, complexity tradeoffs.",
-    },
-    {
-        "code": "cs.CE",
-        "name": "Computational Engineering, Finance, and Science",
-        "description": "Applications of computer science to mathematical modeling of complex systems.",
-    },
-    {
-        "code": "cs.CG",
-        "name": "Computational Geometry",
-        "description": "Roughly includes material in ACM Subject Classes I.3.5 and F.2.2.",
-    },
-    {
-        "code": "cs.CL",
-        "name": "Computation and Language",
-        "description": "Covers natural language processing. Roughly includes material in ACM Subject Class I.2.7.",
-    },
-    {
-        "code": "cs.CR",
-        "name": "Cryptography and Security",
-        "description": "Covers all areas of cryptography and security including authentication, public key cryptosystems.",
-    },
-    {
-        "code": "cs.CV",
-        "name": "Computer Vision and Pattern Recognition",
-        "description": "Covers image processing, computer vision, pattern recognition, and scene understanding.",
-    },
-    {
-        "code": "cs.CY",
-        "name": "Computers and Society",
-        "description": "Covers impact of computers on society, computer ethics, information technology and public policy.",
-    },
-    {
-        "code": "cs.DB",
-        "name": "Databases",
-        "description": "Covers database management, datamining, and data processing.",
-    },
-    {
-        "code": "cs.DC",
-        "name": "Distributed, Parallel, and Cluster Computing",
-        "description": "Covers fault-tolerance, distributed algorithms, stabilization, load balancing.",
-    },
-    {
-        "code": "cs.DL",
-        "name": "Digital Libraries",
-        "description": "Covers all aspects of the digital library design and document creation.",
-    },
-    {
-        "code": "cs.DM",
-        "name": "Discrete Mathematics",
-        "description": "Covers combinatorics, graph theory, applications of probability.",
-    },
-    {
-        "code": "cs.DS",
-        "name": "Data Structures and Algorithms",
-        "description": "Covers data structures and analysis of algorithms.",
-    },
-    {
-        "code": "cs.ET",
-        "name": "Emerging Technologies",
-        "description": "Covers alternatives to silicon CMOS-based technologies.",
-    },
-    {
-        "code": "cs.FL",
-        "name": "Formal Languages and Automata Theory",
-        "description": "Covers automata theory, formal language theory, grammars.",
-    },
-    {
-        "code": "cs.GL",
-        "name": "General Literature",
-        "description": "Covers introductory material, survey material, predictions of future trends.",
-    },
-    {
-        "code": "cs.GR",
-        "name": "Graphics",
-        "description": "Covers all aspects of computer graphics.",
-    },
-    {
-        "code": "cs.GT",
-        "name": "Computer Science and Game Theory",
-        "description": "Covers intersection of computer science and game theory.",
-    },
-    {
-        "code": "cs.HC",
-        "name": "Human-Computer Interaction",
-        "description": "Covers human factors, user interfaces, and collaborative computing.",
-    },
-    {
-        "code": "cs.IR",
-        "name": "Information Retrieval",
-        "description": "Covers indexing, dictionaries, retrieval, content and analysis.",
-    },
-    {
-        "code": "cs.IT",
-        "name": "Information Theory",
-        "description": "Covers theoretical and experimental aspects of information theory and coding.",
-    },
-    {
-        "code": "cs.LG",
-        "name": "Machine Learning",
-        "description": "Papers on all aspects of machine learning research.",
-    },
-    {
-        "code": "cs.LO",
-        "name": "Logic in Computer Science",
-        "description": "Covers all aspects of logic in computer science, including finite model theory.",
-    },
-    {
-        "code": "cs.MA",
-        "name": "Multiagent Systems",
-        "description": "Covers multiagent systems, distributed artificial intelligence.",
-    },
-    {
-        "code": "cs.MM",
-        "name": "Multimedia",
-        "description": "Covers multimedia information systems and multimedia information representation.",
-    },
-    {
-        "code": "cs.MS",
-        "name": "Mathematical Software",
-        "description": "Roughly includes material in ACM Subject Class G.4.",
-    },
-    {
-        "code": "cs.NA",
-        "name": "Numerical Analysis",
-        "description": "Covers numerical algorithms for problems in analysis and algebra.",
-    },
-    {
-        "code": "cs.NE",
-        "name": "Neural and Evolutionary Computing",
-        "description": "Covers neural networks, genetic algorithms, artificial life.",
-    },
-    {
-        "code": "cs.NI",
-        "name": "Networking and Internet Architecture",
-        "description": "Covers all aspects of computer communication networks.",
-    },
-    {
-        "code": "cs.OH",
-        "name": "Other Computer Science",
-        "description": "For documents that do not fit anywhere else.",
-    },
-    {
-        "code": "cs.OS",
-        "name": "Operating Systems",
-        "description": "Operating systems and related topics.",
-    },
-    {
-        "code": "cs.PF",
-        "name": "Performance",
-        "description": "Covers performance measurement and evaluation, queueing, and simulation.",
-    },
-    {
-        "code": "cs.PL",
-        "name": "Programming Languages",
-        "description": "Covers programming language semantics, language features.",
-    },
-    {
-        "code": "cs.RO",
-        "name": "Robotics",
-        "description": "Roughly includes material in ACM Subject Class I.2.9.",
-    },
-    {
-        "code": "cs.SC",
-        "name": "Symbolic Computation",
-        "description": "Roughly includes material in ACM Subject Class I.1.",
-    },
-    {
-        "code": "cs.SD",
-        "name": "Sound",
-        "description": "Covers all aspects of computing with sound.",
-    },
-    {
-        "code": "cs.SE",
-        "name": "Software Engineering",
-        "description": "Covers design tools, software metrics, testing and debugging.",
-    },
-    {
-        "code": "cs.SI",
-        "name": "Social and Information Networks",
-        "description": "Covers design, analysis, and modeling of social and information networks.",
-    },
-    {
-        "code": "cs.SY",
-        "name": "Systems and Control",
-        "description": "Covers theoretical and experimental research on automatic control systems.",
-    },
-]
+import json
+import os
 
-# Mathematics categories
-MATH_CATEGORIES = [
-    {
-        "code": "math.AC",
-        "name": "Commutative Algebra",
-        "description": "Commutative rings, modules, ideals, homological algebra.",
-    },
-    {
-        "code": "math.AG",
-        "name": "Algebraic Geometry",
-        "description": "Algebraic varieties, stacks, sheaves, schemes, moduli spaces.",
-    },
-    {
-        "code": "math.AP",
-        "name": "Analysis of PDEs",
-        "description": "Existence and uniqueness, boundary value problems, linear and non-linear operators.",
-    },
-    {
-        "code": "math.AT",
-        "name": "Algebraic Topology",
-        "description": "Homotopy theory, homological algebra, algebraic treatments of manifolds.",
-    },
-    {
-        "code": "math.CA",
-        "name": "Classical Analysis and ODEs",
-        "description": "Special functions, orthogonal polynomials, harmonic analysis, ODE's.",
-    },
-    {
-        "code": "math.CO",
-        "name": "Combinatorics",
-        "description": "Discrete mathematics, graph theory, enumeration, combinatorial optimization.",
-    },
-    {
-        "code": "math.CT",
-        "name": "Category Theory",
-        "description": "Enriched categories, topoi, abelian categories, monoidal categories.",
-    },
-    {
-        "code": "math.CV",
-        "name": "Complex Variables",
-        "description": "Holomorphic functions, automorphic group actions and forms.",
-    },
-    {
-        "code": "math.DG",
-        "name": "Differential Geometry",
-        "description": "Complex, contact, Riemannian, pseudo-Riemannian and Finsler geometry.",
-    },
-    {
-        "code": "math.DS",
-        "name": "Dynamical Systems",
-        "description": "Dynamics of differential equations and flows, mechanics.",
-    },
-    {
-        "code": "math.FA",
-        "name": "Functional Analysis",
-        "description": "Banach spaces, function spaces, real functions, integral transforms.",
-    },
-    {
-        "code": "math.GM",
-        "name": "General Mathematics",
-        "description": "Mathematical material of general interest.",
-    },
-    {
-        "code": "math.GN",
-        "name": "General Topology",
-        "description": "Continuum theory, point-set topology, spaces with algebraic structure.",
-    },
-    {
-        "code": "math.GR",
-        "name": "Group Theory",
-        "description": "Finite groups, topological groups, representation theory.",
-    },
-    {
-        "code": "math.GT",
-        "name": "Geometric Topology",
-        "description": "Manifolds, orbifolds, polyhedra, cell complexes, foliations.",
-    },
-    {
-        "code": "math.HO",
-        "name": "History and Overview",
-        "description": "Biographies, philosophy of mathematics, mathematics education.",
-    },
-    {
-        "code": "math.IT",
-        "name": "Information Theory",
-        "description": "Alias for cs.IT. Covers information theory and coding.",
-    },
-    {
-        "code": "math.KT",
-        "name": "K-Theory and Homology",
-        "description": "Algebraic and topological K-theory.",
-    },
-    {
-        "code": "math.LO",
-        "name": "Logic",
-        "description": "Logic, set theory, point-set topology, formal mathematics.",
-    },
-    {
-        "code": "math.MG",
-        "name": "Metric Geometry",
-        "description": "Euclidean, hyperbolic, discrete, convex, coarse geometry.",
-    },
-    {
-        "code": "math.MP",
-        "name": "Mathematical Physics",
-        "description": "Alias for math-ph. Application of mathematics to physics.",
-    },
-    {
-        "code": "math.NA",
-        "name": "Numerical Analysis",
-        "description": "Alias for cs.NA. Numerical algorithms.",
-    },
-    {
-        "code": "math.NT",
-        "name": "Number Theory",
-        "description": "Prime numbers, diophantine equations, analytic number theory.",
-    },
-    {
-        "code": "math.OA",
-        "name": "Operator Algebras",
-        "description": "Algebras of operators on Hilbert space, C^*-algebras.",
-    },
-    {
-        "code": "math.OC",
-        "name": "Optimization and Control",
-        "description": "Operations research, linear programming, control theory.",
-    },
-    {
-        "code": "math.PR",
-        "name": "Probability",
-        "description": "Theory and applications of probability and stochastic processes.",
-    },
-    {
-        "code": "math.QA",
-        "name": "Quantum Algebra",
-        "description": "Quantum groups, skein theories, operadic and diagrammatic algebra.",
-    },
-    {
-        "code": "math.RA",
-        "name": "Rings and Algebras",
-        "description": "Non-commutative rings and algebras, non-associative algebras.",
-    },
-    {
-        "code": "math.RT",
-        "name": "Representation Theory",
-        "description": "Linear representations of algebras and groups, Lie theory.",
-    },
-    {
-        "code": "math.SG",
-        "name": "Symplectic Geometry",
-        "description": "Hamiltonian systems, symplectic flows, classical integrable systems.",
-    },
-    {
-        "code": "math.SP",
-        "name": "Spectral Theory",
-        "description": "Schrodinger operators, operators on manifolds.",
-    },
-    {
-        "code": "math.ST",
-        "name": "Statistics Theory",
-        "description": "Alias for stat.TH. Statistical inference, regression, time series.",
-    },
-]
+_DATA_FILE = os.path.join(os.path.dirname(__file__), "arxiv_categories.json")
 
-# Physics categories
-PHYSICS_CATEGORIES = [
-    {
-        "code": "physics.acc-ph",
-        "name": "Accelerator Physics",
-        "description": "Accelerator theory and simulation.",
-    },
-    {
-        "code": "physics.ao-ph",
-        "name": "Atmospheric and Oceanic Physics",
-        "description": "Atmospheric and oceanic physics and physical chemistry.",
-    },
-    {
-        "code": "physics.app-ph",
-        "name": "Applied Physics",
-        "description": "Applications of physics to new technology.",
-    },
-    {
-        "code": "physics.atm-clus",
-        "name": "Atomic and Molecular Clusters",
-        "description": "Atomic and molecular clusters, nanoparticles.",
-    },
-    {
-        "code": "physics.atom-ph",
-        "name": "Atomic Physics",
-        "description": "Atomic and molecular structure, spectra, collisions.",
-    },
-    {
-        "code": "physics.bio-ph",
-        "name": "Biological Physics",
-        "description": "Molecular biophysics, cellular biophysics, neurological biophysics.",
-    },
-    {
-        "code": "physics.chem-ph",
-        "name": "Chemical Physics",
-        "description": "Experimental, computational, and theoretical physics of atoms, molecules.",
-    },
-    {
-        "code": "physics.class-ph",
-        "name": "Classical Physics",
-        "description": "Newtonian mechanics, classical field theories.",
-    },
-    {
-        "code": "physics.comp-ph",
-        "name": "Computational Physics",
-        "description": "Computational methods in physics and related fields.",
-    },
-    {
-        "code": "physics.data-an",
-        "name": "Data Analysis, Statistics and Probability",
-        "description": "Methods, software and hardware for physics data analysis.",
-    },
-    {
-        "code": "physics.ed-ph",
-        "name": "Physics Education",
-        "description": "Report of results of a research study within physics education.",
-    },
-    {
-        "code": "physics.flu-dyn",
-        "name": "Fluid Dynamics",
-        "description": "Turbulence, instabilities, incompressible/compressible flows.",
-    },
-    {
-        "code": "physics.gen-ph",
-        "name": "General Physics",
-        "description": "General physics topics.",
-    },
-    {
-        "code": "physics.geo-ph",
-        "name": "Geophysics",
-        "description": "Atmospheric physics. Biogeosciences. Computational geophysics.",
-    },
-    {
-        "code": "physics.hist-ph",
-        "name": "History and Philosophy of Physics",
-        "description": "History and philosophy of physics, astronomy, and cosmology.",
-    },
-    {
-        "code": "physics.ins-det",
-        "name": "Instrumentation and Detectors",
-        "description": "Instrumentation and detectors for research.",
-    },
-    {
-        "code": "physics.med-ph",
-        "name": "Medical Physics",
-        "description": "Radiation therapy. Diagnostic radiology. Biomedical imaging.",
-    },
-    {
-        "code": "physics.optics",
-        "name": "Optics",
-        "description": "Adaptive optics, astronomical optics, atmospheric optics.",
-    },
-    {
-        "code": "physics.plasm-ph",
-        "name": "Plasma Physics",
-        "description": "Fundamental plasma physics, magnetic confinement fusion.",
-    },
-    {
-        "code": "physics.pop-ph",
-        "name": "Popular Physics",
-        "description": "Popular physics topics.",
-    },
-    {
-        "code": "physics.soc-ph",
-        "name": "Physics and Society",
-        "description": "Structure, dynamics and collective behavior of societies.",
-    },
-    {
-        "code": "physics.space-ph",
-        "name": "Space Physics",
-        "description": "Heliophysics, solar physics, heliospheric physics.",
-    },
-]
 
-# Statistics categories
-STAT_CATEGORIES = [
-    {
-        "code": "stat.AP",
-        "name": "Applications",
-        "description": "Biology, Education, Epidemiology, Engineering, Environmental Sciences.",
-    },
-    {
-        "code": "stat.CO",
-        "name": "Computation",
-        "description": "Algorithms, Simulation, Visualization.",
-    },
-    {
-        "code": "stat.ME",
-        "name": "Methodology",
-        "description": "Design, Surveys, Model Selection, Multiple Testing.",
-    },
-    {
-        "code": "stat.ML",
-        "name": "Machine Learning",
-        "description": "Machine learning papers with statistical or theoretical grounding.",
-    },
-    {
-        "code": "stat.OT",
-        "name": "Other Statistics",
-        "description": "Work in statistics that does not fit into the other stat classifications.",
-    },
-    {
-        "code": "stat.TH",
-        "name": "Statistics Theory",
-        "description": "Alias for math.ST. Asymptotics, Bayesian Inference, Decision Theory.",
-    },
-]
-
-# Quantitative Biology categories
-QBIO_CATEGORIES = [
-    {
-        "code": "q-bio.BM",
-        "name": "Biomolecules",
-        "description": "DNA, RNA, proteins, lipids; molecular structures and folding kinetics.",
-    },
-    {
-        "code": "q-bio.CB",
-        "name": "Cell Behavior",
-        "description": "Cell-cell signaling and interaction; morphogenesis and development.",
-    },
-    {
-        "code": "q-bio.GN",
-        "name": "Genomics",
-        "description": "DNA sequencing and assembly; gene and motif finding.",
-    },
-    {
-        "code": "q-bio.MN",
-        "name": "Molecular Networks",
-        "description": "Gene regulation, signal transduction, proteomics, metabolomics.",
-    },
-    {
-        "code": "q-bio.NC",
-        "name": "Neurons and Cognition",
-        "description": "Synapse, cortex, neuronal dynamics, neural network.",
-    },
-    {
-        "code": "q-bio.OT",
-        "name": "Other Quantitative Biology",
-        "description": "Work in quantitative biology that does not fit elsewhere.",
-    },
-    {
-        "code": "q-bio.PE",
-        "name": "Populations and Evolution",
-        "description": "Population dynamics, spatio-temporal and epidemiological models.",
-    },
-    {
-        "code": "q-bio.QM",
-        "name": "Quantitative Methods",
-        "description": "All experimental, numerical, statistical and mathematical contributions.",
-    },
-    {
-        "code": "q-bio.SC",
-        "name": "Subcellular Processes",
-        "description": "Assembly and control of subcellular structures.",
-    },
-    {
-        "code": "q-bio.TO",
-        "name": "Tissues and Organs",
-        "description": "Blood flow in vessels, biomechanics of bones.",
-    },
-]
+def _load_categories():
+    with open(_DATA_FILE, "r", encoding="utf-8") as f:
+        return json.load(f)
 
 
 def get_all_categories():
     """Return all arXiv categories as a single list."""
-    return (
-        CS_CATEGORIES
-        + MATH_CATEGORIES
-        + PHYSICS_CATEGORIES
-        + STAT_CATEGORIES
-        + QBIO_CATEGORIES
-    )
+    data = _load_categories()
+    result = []
+    for section in data.values():
+        result.extend(section)
+    return result
+
+
+# Legacy named lists (loaded on demand to avoid startup cost)
+def _get_section(key):
+    return _load_categories()[key]
+
+
+CS_CATEGORIES = None
+MATH_CATEGORIES = None
+PHYSICS_CATEGORIES = None
+STAT_CATEGORIES = None
+QBIO_CATEGORIES = None
+
+
+def __getattr__(name):
+    _map = {
+        "CS_CATEGORIES": "cs",
+        "MATH_CATEGORIES": "math",
+        "PHYSICS_CATEGORIES": "physics",
+        "STAT_CATEGORIES": "stat",
+        "QBIO_CATEGORIES": "q-bio",
+    }
+    if name in _map:
+        return _get_section(_map[name])
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
 
 
 # EOF
