+++ apps/writer_app/static/writer_app/css/components/status-lamp.css
@@ -13,6 +13,7 @@
   display: flex;
   align-items: center;
   gap: 0.25rem;
+  cursor: pointer;
 }
 
 /* Status indicator label text */
