+++ apps/writer_app/static/writer_app/css/shared/index-ui/00-index-ui.css
@@ -4,3 +4,16 @@
  * =============================================================================
  */
 
+/* Writer workspace layout */
+.writer-workspace {
+  flex: 1;
+  display: flex;
+  height: 100%;
+  overflow: hidden;
+}
+
+.writer-container {
+  flex: 1;
+  min-width: 0;
+  overflow: hidden;
+}
