+++ apps/writer_app/templates/writer_app/index_partials/details_panel.html
@@ -68,9 +68,7 @@
                         onclick="if(window.handleFullCompileClick){window.handleFullCompileClick()}">
                     <i class="fas fa-rocket me-1"></i> Compile Full
                 </button>
-                <button id="details-compile-btn"
-                        class="btn btn-primary btn-sm w-100 mb-2"
-                        style="display: none">
+                <button id="details-compile-btn" class="btn btn-primary btn-sm w-100 mb-2 hidden">
                     <i class="fas fa-rocket me-1"></i> Compile Full
                 </button>
                 <div id="details-full-log" class="details-compilation-log">Compilation log will be shown here</div>
@@ -82,7 +80,7 @@
              title="Import from / Export to Overleaf">
             <div class="mig-card__header">
                 <h6 class="mig-card__title">
-                    <i class="fas fa-leaf" style="color:#47a141;"></i> &#x2194; Overleaf
+                    <i class="fas fa-leaf icon-overleaf"></i> &#x2194; Overleaf
                     <span class="badge badge-wip badge-sm">WIP</span>
                 </h6>
                 <i class="fas fa-chevron-down mig-card__chevron"></i>
@@ -107,7 +105,7 @@
                     <i class="fas fa-file-import"></i> Import ZIP
                 </button>
                 <div id="overleaf-import-status" class="mig-card__status" style="display:none;"></div>
-                <hr style="border-color:var(--workspace-border-default);margin:8px 0;" />
+                <hr class="mig-card__separator" />
                 <button id="overleaf-export-btn" class="mig-card__btn">
                     <i class="fas fa-file-export"></i> Export to Overleaf ZIP
                 </button>
@@ -120,7 +118,7 @@
              title="Import from / Export to Prism (OpenAI)">
             <div class="mig-card__header">
                 <h6 class="mig-card__title">
-                    <i class="fas fa-atom" style="color:#10a37f;"></i> &#x2194; Prism (OpenAI)
+                    <i class="fas fa-atom icon-prism"></i> &#x2194; Prism (OpenAI)
                     <span class="badge badge-wip badge-sm">WIP</span>
                 </h6>
                 <i class="fas fa-chevron-down mig-card__chevron"></i>
@@ -131,7 +129,7 @@
                     Use the following workflows to transfer your work.
                 </p>
                 <label class="mig-card__label">
-                    <i class="fas fa-sign-in-alt" style="color:#10a37f;"></i> Import from Prism
+                    <i class="fas fa-sign-in-alt" class="icon-prism"></i> Import from Prism
                 </label>
                 <ol class="mig-card__steps">
                     <li>
@@ -143,7 +141,7 @@
                     <li>Or export BibTeX and use Scholar &rarr; Zotero Import</li>
                 </ol>
                 <label class="mig-card__label">
-                    <i class="fas fa-sign-out-alt" style="color:#10a37f;"></i> Export to Prism
+                    <i class="fas fa-sign-out-alt" class="icon-prism"></i> Export to Prism
                 </label>
                 <ol class="mig-card__steps">
                     <li>
@@ -151,7 +149,7 @@
                     </li>
                     <li>In Prism, import the ZIP as a new project</li>
                 </ol>
-                <p class="mig-card__hint" style="margin-bottom:0.5rem;">
+                <p class="mig-card__hint mb-2">
                     <i class="fas fa-info-circle"></i>
                     No vendor lock-in — freely move between SciTeX, Prism, Overleaf, and Zotero.
                 </p>
