+++ apps/writer_app/templates/writer_app/writer_partial.html
@@ -1,87 +1,6 @@
 {% load static vite tree_preseed %}
 
-<!-- Writer CSS — moved to <head> by unified workspace JS -->
-<link rel="stylesheet"
-      href="{% static 'shared/css/components/workspace-files-tree.css' %}?v=20251202d" />
-<link rel="stylesheet"
-      href="{% static 'shared/css/components/workspace-sidebar.css' %}" />
-<link rel="stylesheet" href="{% static 'shared/css/components/panel-resizer.css' %}" />
-<link rel="stylesheet"
-      href="{% static 'shared/css/components/file-tabs.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="{% static 'writer_app/css/editor/tex-view-main.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="{% static 'writer_app/css/editor/index-editor-panels.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="{% static 'writer_app/css/editor/pdf-view-main.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="{% static 'writer_app/css/components/pdfjs-viewer.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="{% static 'shared/css/components/alerts.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="{% static 'shared/css/components/terminal-log.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="{% static 'writer_app/css/shared/index-ui/00-index-ui.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="{% static 'writer_app/css/shared/index-ui/01-panel-header.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="{% static 'writer_app/css/shared/index-ui/02-latex-panel.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="{% static 'writer_app/css/shared/index-ui/03-preview-panel.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="{% static 'writer_app/css/shared/index-ui/04-toolbar.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="{% static 'writer_app/css/shared/index-ui/05-compilation-output.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="{% static 'writer_app/css/shared/index-ui/06-split-view.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="{% static 'writer_app/css/shared/index-ui/07-panel-resizer.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="{% static 'writer_app/css/shared/index-ui/08-save-status.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="{% static 'writer_app/css/shared/index-ui/09-button-styling.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="{% static 'writer_app/css/shared/index-ui/10-dark-mode.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="{% static 'writer_app/css/shared/index-ui/11-responsive.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="{% static 'writer_app/css/shared/index-ui/12-initialization-prompt.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="{% static 'writer_app/css/shared/index-ui/13-accessibility.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="{% static 'writer_app/css/shared/index-ui/15-details-panel.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="{% static 'writer_app/css/components/status-lamp.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="{% static 'writer_app/css/components/shortcuts-modal.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="{% static 'writer_app/css/components/sidebar-controls.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="{% static 'writer_app/css/editor/main-editor-inline.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="{% static 'writer_app/css/editor/file-tabs.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="{% static 'writer_app/css/editor/citations-panel/index.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="{% static 'writer_app/css/editor/figures-panel/index.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="{% static 'writer_app/css/editor/tables-panel.css' %}?v={{ build_id }}" />
-<link rel="stylesheet"
-      href="https://cdnjs.cloudflare.com/ajax/libs/monaco-editor/0.45.0/min/vs/editor/editor.main.min.css" />
-<style>
-    .writer-workspace {
-        flex: 1;
-        display: flex;
-        height: 100%;
-        overflow: hidden;
-    }
-
-    .writer-container {
-        flex: 1;
-        min-width: 0;
-        overflow: hidden;
-    }
-</style>
+{% include "writer_app/index_partials/_writer_css.html" %}
 {% if needs_project_creation %}
     <div class="container py-5">
         <div class="row justify-content-center">
